test: add app view test helpers
